requirements.txt
streamlit>=1.32.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.2
openai>=1.12.0
python-dotenv>=1.0.0
